requirements.txt
openai>=1.0.0
pyyaml>=6.0.1
rich>=13.0.0
matplotlib>=3.8.0
numpy>=1.24.0
pandas>=2.0.0
